feat: separate generic handwriting and maths single answer task
