+++ src/components/questionStructure/Task/constants.ts
@@ -5,5 +5,6 @@ export enum TaskType {
   FLAG = 'FLAG',
   MULTIPLE_CHOICE_SELECT_ONE = 'MULTIPLE_CHOICE_SELECT_ONE',
   MULTIPLE_CHOICE_SELECT_SEVERAL = 'MULTIPLE_CHOICE_SELECT_SEVERAL',
-  PROCESSED_HANDWRITING = 'PROCESSED_HANDWRITING',
+  MATHS_SINGLE_ANSWER = 'MATHS_SINGLE_ANSWER',
+  HANDWRITING = 'HANDWRITING',
 }
