+++ src/components/questionStructure/Task/variants/handwritingTasks/index.scss
