+++ src/components/questionStructure/Task/variants/handwritingTasks/types.ts
@@ -1,5 +1,6 @@
 import { ExcalidrawElement } from '@excalidraw/excalidraw/types/element/types'
 import { AppState } from '@excalidraw/excalidraw/types/types'
+import React from 'react'
 
 export interface HandwritingAnswer {
   raw: {
@@ -10,4 +11,10 @@ export interface HandwritingAnswer {
 
 export interface MathsSingleAnswer extends HandwritingAnswer {
   latex: string
+}
+
+export interface GenericHandwritingEditorProps<T> {
+  answer?: T
+  onAnswerChange: (value: T) => void
+  children?: React.ReactNode
 }
