+++ src/components/questionStructure/Task/variants/handwritingTasks/variants/ViewOnlyCanvas.tsx
