+++ src/components/questionStructure/Task/variants/handwritingTasks/variants/handwritingEditor.css
